requirements.txt
docopt
GitPython
py
git+https://github.com/tbarron/tbx

